Uodatating plotting, run_all_over

+++ ood_scores/ood_extractors.py
@@ -32,9 +32,12 @@ class OODScoresExtractor(OODScores):
             if not path.exists(output_dir):
                 makedirs(output_dir)
             file_dir = f"{output_dir}/ood_scores_fit_samples_b{batch_size}_{self.fit_ds_name}_using_checkpoint_{checkpoint_number}.pth"
-            ood_scores_fit_samples = self.run_ood_scores(dataset, batch_size, means, variances, num_samples)
-            torch.save(ood_scores_fit_samples, file_dir)
-            print(f"fit_ood_scores of {self.fit_ds_name} is saved in\n {file_dir}")
+            if not path.exists(file_dir):
+                ood_scores_fit_samples = self.run_ood_scores(dataset, batch_size, means, variances, num_samples)
+                torch.save(ood_scores_fit_samples, file_dir)
+                print(f"fit_ood_scores of {self.fit_ds_name} is saved in\n {file_dir}")
+            else:
+                print(f"fit_ood_scores of {self.fit_ds_name} already exists in\n {file_dir}")
         else:
             # Compute OOD scores for Test dataset
             output_dir = self.path_creator.model_dataset_path(self.model.__class__.__name__.lower(), self.fit_ds_name)
@@ -42,10 +45,12 @@ class OODScoresExtractor(OODScores):
             if not path.exists(output_dir):
                 makedirs(output_dir)
             file_dir = f"{output_dir}/ood_scores_test_samples_b{batch_size}_{ds_name}_on_{self.fit_ds_name}_using_checkpoint_{checkpoint_number}.pth"
-            ood_scores_test_samples = self.run_ood_scores(dataset, batch_size, means, variances, num_samples)
-            torch.save(ood_scores_test_samples, file_dir)
-            print(f"Test_ood_scores of {ds_name} is saved in\n {file_dir}")
-
+            if not path.exists(file_dir):
+                ood_scores_test_samples = self.run_ood_scores(dataset, batch_size, means, variances, num_samples)
+                torch.save(ood_scores_test_samples, file_dir)
+                print(f"Test_ood_scores of {ds_name} is saved in\n {file_dir}")
+            else:
+                print(f"Test_ood_scores of {ds_name} already exists in\n {file_dir}")
 
     def run_ood_scores(self, ds, b_size, means, variances, num_samples):
         """Compute OOD scores for each sample in the batch"""
